got drop down and working on organizing individual posts
