+++ homeaway/package.json
@@ -4,9 +4,10 @@
   "private": true,
   "dependencies": {
     "react": "^16.9.0",
+    "react-bootstrap": "^1.0.0-beta.14",
     "react-dom": "^16.9.0",
-    "react-scripts": "3.1.2",
-    "react-router-dom": "latest"
+    "react-router-dom": "latest",
+    "react-scripts": "3.1.2"
   },
   "scripts": {
     "start": "react-scripts start",
